+++ CHANGELOG.md
@@ -1,6 +1,6 @@
 # Change Log
 
-All notable changes to the "WeaveContext" extension will be documented in this file.
+All notable changes to the "Weave Context" extension will be documented in this file.
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
@@ -11,7 +11,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 
-- Initial release of WeaveContext extension
+- Initial release of Weave Context extension
 
 [Unreleased]: https://github.com/ManuelGil/vscode-weave-context/compare/v1.0.0...HEAD
 [1.0.0]: https://github.com/ManuelGil/vscode-weave-context/releases/tag/v1.0.0